requirements.txt
# Core ML/Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
imbalanced-learn>=0.11.0

# Workflow Orchestration
prefect>=2.14.0

# Experiment Tracking
mlflow>=2.9.0

# Drift Detection & Monitoring
evidently>=0.4.22

# Data Versioning (optional for now)
# dvc>=3.30.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Processing
scipy>=1.11.0

# API (for future deployment)
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
flake8>=6.1.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Logging
loguru>=0.7.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.27.0